api/requirements.txt
albumentations==1.3.0
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
fastapi==0.92.0
fonttools==4.38.0
h11==0.14.0
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
motor==3.1.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==23.0
Pillow==9.4.0
pydantic==1.10.5
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
threadpoolctl==3.1.0
tifffile==2023.2.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
zipp==3.14.0
